Minor cleanup
* renamed IDRequest to IdRequest
* extracted IdRequest into lib/id_request.rb
* renamed lib/clientrequest.rb to lib/client_request.rb (ruby convention)
* added spec for IdRequest
