+++ lib/client_request.rb
@@ -8,7 +8,6 @@ $LOAD_PATH << '..'
 require 'KibanaConfig'
 require 'compat'
 
-
 =begin
 = Class: ClientRequest
   Creates an object out of the hash passed by a client
@@ -64,33 +63,5 @@ class ClientRequest
     def hash(request)
       Base64.encode64(JSON.generate(request))
     end
-
-  end
-
-end
-
-=begin
-= Class: IDRequest
-  Creates an client request for a specific ID
-== Parameters:
-  id::    ID of log
-  index:: Index ID is found in
-=end
-class IDRequest
-  attr_accessor :request
-  def initialize(id,index)
-    @request = {
-      "id"        => "#{id}",
-      "index"     => index,
-      "timeframe" => "900",
-      "mode"      => "id",
-      "fields"    => '',
-      "offset"    => 0,
-    }
-  end
-
-  def hash
-    ClientRequest.hash(@request)
   end
 end
-
