+++ lib/id_request.rb
@@ -1,0 +1,28 @@
+require "client_request"
+
+=begin
+= Class: IdRequest
+  Creates an client request for a specific Id
+== Parameters:
+  id::    Id of log
+  index:: Index Id is found in
+=end
+class IdRequest
+  attr_accessor :request
+
+  def initialize(id,index)
+    @request = {
+      "id"        => "#{id}",
+      "index"     => index,
+      "timeframe" => "900",
+      "mode"      => "id",
+      "fields"    => '',
+      "offset"    => 0,
+    }
+  end
+
+  def hash
+    ClientRequest.hash(@request)
+  end
+end
+
